Refactor and consolidate quest-related services into `QuestEngine`; remove deprecated classes and adapt usages across modules.
