+++ composeApp/src/commonMain/kotlin/io/yavero/aterna/domain/quest/economy/RewardLedgerEntry.kt
@@ -1,4 +1,4 @@
-package io.yavero.aterna.domain.service.quest
+package io.yavero.aterna.domain.quest.economy
 
 data class RewardLedgerEntry(
     val eventIdx: Int,
