+++ composeApp/src/commonMain/kotlin/io/yavero/aterna/domain/quest/engine/CompleteResult.kt
@@ -1,4 +1,4 @@
-package io.yavero.aterna.domain.service.quest
+package io.yavero.aterna.domain.quest.engine
 
 import io.yavero.aterna.domain.model.Hero
 import io.yavero.aterna.domain.model.Quest
@@ -6,10 +6,11 @@ import io.yavero.aterna.domain.model.QuestLoot
 import io.yavero.aterna.features.quest.presentation.QuestEffect
 
 
-data class RetreatResult(
+data class CompleteResult(
     val quest: Quest,
     val updatedHero: Hero,
-    val bankedLoot: QuestLoot?,
-    val curseApplied: Boolean,
+    val loot: QuestLoot,
+    val leveledUpTo: Int?,
+    val newItemIds: Set<String> = emptySet(),
     val uiEffects: List<QuestEffect> = emptyList(),
 )
