+++ composeApp/src/commonMain/kotlin/io/yavero/aterna/domain/quest/engine/RetreatResult.kt
@@ -1,4 +1,4 @@
-package io.yavero.aterna.domain.service.quest
+package io.yavero.aterna.domain.quest.engine
 
 import io.yavero.aterna.domain.model.Hero
 import io.yavero.aterna.domain.model.Quest
@@ -6,11 +6,10 @@ import io.yavero.aterna.domain.model.QuestLoot
 import io.yavero.aterna.features.quest.presentation.QuestEffect
 
 
-data class CompleteResult(
+data class RetreatResult(
     val quest: Quest,
     val updatedHero: Hero,
-    val loot: QuestLoot,
-    val leveledUpTo: Int?,
-    val newItemIds: Set<String> = emptySet(),
+    val bankedLoot: QuestLoot?,
+    val curseApplied: Boolean,
     val uiEffects: List<QuestEffect> = emptyList(),
 )
