+++ composeApp/src/commonMain/kotlin/io/yavero/aterna/domain/quest/engine/StartResult.kt
@@ -1,9 +1,10 @@
-@file:OptIn(kotlin.time.ExperimentalTime::class)
+@file:OptIn(ExperimentalTime::class)
 
-package io.yavero.aterna.domain.service.quest
+package io.yavero.aterna.domain.quest.engine
 
 import io.yavero.aterna.domain.model.Quest
 import io.yavero.aterna.features.quest.presentation.QuestEffect
+import kotlin.time.ExperimentalTime
 import kotlin.time.Instant
