+++ composeApp/src/commonMain/kotlin/io/yavero/aterna/domain/quest/ticker/DefaultTicker.kt
@@ -1,6 +1,6 @@
-@file:OptIn(kotlin.time.ExperimentalTime::class)
+@file:OptIn(ExperimentalTime::class)
 
-package io.yavero.aterna.domain.service.ticker
+package io.yavero.aterna.domain.quest.ticker
 
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.delay
@@ -9,6 +9,7 @@ import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.flow.shareIn
 import kotlin.time.Clock
+import kotlin.time.ExperimentalTime
 import kotlin.time.Instant
 
 class DefaultTicker(scope: CoroutineScope) : Ticker {
