+++ composeApp/src/commonMain/kotlin/io/yavero/aterna/domain/service/quest/QuestActionService.kt
@@ -1,6 +1,9 @@
 package io.yavero.aterna.domain.service.quest
 
 import io.yavero.aterna.domain.model.ClassType
+import io.yavero.aterna.domain.quest.engine.CompleteResult
+import io.yavero.aterna.domain.quest.engine.RetreatResult
+import io.yavero.aterna.domain.quest.engine.StartResult
 
 interface QuestActionService {
     suspend fun start(durationMinutes: Int, classType: ClassType): StartResult
