--- composeApp/src/commonMain/kotlin/io/yavero/aterna/domain/util/FixedIntervalBankingStrategy.kt
@@ -1,12 +1,0 @@
-package io.yavero.aterna.domain.util
-
-class FixedIntervalBankingStrategy(
-    private val intervalMinutes: Int = 10
-) : RewardBankingStrategy {
-
-    override fun bankedElapsedMs(elapsedMs: Long): Long {
-        val step = intervalMinutes * 60_000L
-        if (step <= 0) return 0
-        return (elapsedMs / step) * step
-    }
-}
