--- composeApp/src/commonMain/kotlin/io/yavero/aterna/domain/util/RewardBanking.kt
@@ -1,6 +1,0 @@
-package io.yavero.aterna.domain.util
-
-
-interface RewardBankingStrategy {
-    fun bankedElapsedMs(elapsedMs: Long): Long
-}
